requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database dependencies
motor==3.3.2
pymongo==4.6.0

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# CORS middleware
python-multipart==0.0.6

# Environment and configuration
python-dotenv==1.0.0

# Additional utilities
email-validator==2.1.0

python-jose[cryptography]==45.0.5